+++ packages/robusta_runner/CHANGELOG.md
@@ -1,4 +1,17 @@
 ## 0.2.0
+
+> Note: This release has breaking changes.
+
+ - hello
+
+ - **FEAT**: add opt out options for event & logger aware callbacks. ([1edc88d2](https://github.com/covalab/robusta/commit/1edc88d23db559ebcc7a4c6906f5a02be599caae))
+ - **FEAT**: add implementing callback extension configurator. ([39fcbcce](https://github.com/covalab/robusta/commit/39fcbccef7ef6f13fc6f14d611d60c83e48ccde0))
+ - **FEAT**: add dependence extension. ([1ba7dda2](https://github.com/covalab/robusta/commit/1ba7dda2cec31c34f9f468fefde0003eecd200d8))
+ - **BREAKING** **FEAT**: replace boots arg with defineBoot. ([1067f066](https://github.com/covalab/robusta/commit/1067f0661da855be2e22cfe394460ba739d7c0ff))
+ - **BREAKING** **FEAT**: improve error handling. ([2abb2014](https://github.com/covalab/robusta/commit/2abb2014b0a1e86e395ce079ada13c8e417eed11))
+ - **BREAKING** **FEAT**: normalize builtin extensions. ([09c14b61](https://github.com/covalab/robusta/commit/09c14b6175f6a49d0597d99b1d8cc317c5e62c4a))
+
+## 0.2.0
 - Add dependence extension.
 - BC break: normalize `EventExtension` and `ImplementingCallbackExtension`.
 - BC break: replace `ExceptionEvent` with `ErrorEvent`.